Improved the release:prepare-packages script by compiling everything to the "release/" directory and then, moving it to the "release/npm" directory to use the same directory pattern as local packages. Added a confirmation question when the "release/" directory is not empty.

+++ scripts/release/preparepackages.js
@@ -79,14 +79,40 @@ const tasks = new Listr( [
 		}
 	},
 	{
-		title: 'Create the release directory.',
-		task: async () => {
-			await fs.emptyDir( RELEASE_DIRECTORY );
+		title: 'Check the release directory.',
+		task: async ( ctx, task ) => {
+			const isAvailable = await fs.exists( RELEASE_DIRECTORY );
+
+			if ( !isAvailable ) {
+				return fs.ensureDir( RELEASE_DIRECTORY );
+			}
+
+			const isEmpty = ( await fs.readdir( RELEASE_DIRECTORY ) ).length === 0;
+
+			if ( isEmpty ) {
+				return Promise.resolve();
+			}
+
+			// Do not ask when running on CI.
+			if ( cliArguments.ci ) {
+				return fs.emptyDir( RELEASE_DIRECTORY );
+			}
+
+			const shouldContinue = await task.prompt( {
+				type: 'Confirm',
+				message: 'The release directory must be empty. Continue and remove all files?'
+			} );
+
+			if ( !shouldContinue ) {
+				return Promise.reject( 'Aborting as requested.' );
+			}
+
+			return fs.emptyDir( RELEASE_DIRECTORY );
 		}
 	},
 	{
 		title: 'Preparation phase.',
-		task: ( _, task ) => {
+		task: ( ctx, task ) => {
 			return task.newListr( [
 				{
 					title: 'Updating "version" value.',
@@ -130,7 +156,7 @@ const tasks = new Listr( [
 	},
 	{
 		title: 'Compilation phase.',
-		task: ( _, task ) => {
+		task: ( ctx, task ) => {
 			return task.newListr( [
 				{
 					title: 'Preparing the "ckeditor5" package files.',
@@ -178,7 +204,7 @@ const tasks = new Listr( [
 					title: 'Copying CKEditor 5 packages to the release directory.',
 					task: () => {
 						return releaseTools.prepareRepository( {
-							outputDirectory: RELEASE_NPM_DIRECTORY,
+							outputDirectory: RELEASE_DIRECTORY,
 							packagesDirectory: PACKAGES_DIRECTORY,
 							rootPackageJson: getCKEditor5PackageJson(),
 							packagesToCopy: cliArguments.packages
@@ -189,7 +215,7 @@ const tasks = new Listr( [
 					title: 'Updating entries in `package.json`.',
 					task: ( ctx, task ) => {
 						return releaseTools.executeInParallel( {
-							packagesDirectory: RELEASE_NPM_DIRECTORY,
+							packagesDirectory: RELEASE_DIRECTORY,
 							listrTask: task,
 							taskToExecute: updatePackageEntryPoint,
 							concurrency: cliArguments.concurrency
@@ -200,7 +226,7 @@ const tasks = new Listr( [
 					title: 'Preparing DLL builds.',
 					task: ( ctx, task ) => {
 						return releaseTools.executeInParallel( {
-							packagesDirectory: RELEASE_NPM_DIRECTORY,
+							packagesDirectory: RELEASE_DIRECTORY,
 							listrTask: task,
 							taskToExecute: prepareDllBuildsCallback,
 							concurrency: cliArguments.concurrency,
@@ -208,6 +234,21 @@ const tasks = new Listr( [
 								RELEASE_CDN_DIRECTORY
 							}
 						} );
+					}
+				},
+				{
+					title: 'Moving packages to npm release directory.',
+					task: async () => {
+						const movePromises = ( await fs.readdir( RELEASE_DIRECTORY ) )
+							.filter( packageSlug => packageSlug.startsWith( 'ckeditor5' ) )
+							.map( packageSlug => {
+								return fs.move(
+									upath.join( RELEASE_DIRECTORY, packageSlug ),
+									upath.join( RELEASE_NPM_DIRECTORY, packageSlug )
+								);
+							} );
+
+						return Promise.all( movePromises );
 					}
 				},
 				{
@@ -246,7 +287,7 @@ const tasks = new Listr( [
 	},
 	{
 		title: 'Clean up phase.',
-		task: ( _, task ) => {
+		task: ( ctx, task ) => {
 			return task.newListr( [
 				{
 					title: 'Removing files that will not be published to npm.',
